30/04

+++ sources/structures/Edge.h
@@ -51,7 +51,7 @@ public:
 		return child;
 	}
 
-	long getId() const {
+	int getId() const {
 		return id;
 	}
